+++ aiida/work/job_processes.py
@@ -18,6 +18,7 @@ from aiida.common.lang import override
 from aiida.daemon import execmanager
 from aiida.orm.calculation.job import JobCalculation
 from aiida.scheduler.datastructures import job_states
+from aiida.work.process_spec import DictSchema
 
 from . import processes
 from . import utils
@@ -107,7 +108,7 @@ class JobProcess(processes.Process):
                 'prepend_text': unicode,
                 'append_text': unicode,
             }
-            spec.input(cls.OPTIONS_INPUT_LABEL, validator=processes.DictSchema(options), non_db=True)
+            spec.input(cls.OPTIONS_INPUT_LABEL, validator=DictSchema(options), non_db=True)
 
             # Inputs from use methods
             for key, use_method in calc_class._use_methods.iteritems():
