+++ aiida/work/ports.py
@@ -1,0 +1,29 @@
+# -*- coding: utf-8 -*-
+from plum import port
+
+
+class WithNonDb(object):
+    """
+    A mixin that adds support to a port to flag a that should not be stored
+    in the database using the non_db=True flag.
+
+    The mixins have to go before the main port class in the superclass order
+    to make sure the mixin has the chance to strip out the non_db keyword.
+    """
+
+    def __init__(self, *args, **kwargs):
+        non_db = kwargs.pop('non_db', False)
+        super(WithNonDb, self).__init__(*args, **kwargs)
+        self._non_db = non_db
+
+    @property
+    def non_db(self):
+        return self._non_db
+
+
+class InputPort(WithNonDb, port.InputPort):
+    pass
+
+
+class PortNamespace(WithNonDb, port.PortNamespace):
+    pass
